feature: Add API service base
Implements: #1
