+++ api/src/main/java/gaiduchek/maksym/api/exceptions/AccessException.java
@@ -1,0 +1,10 @@
+package gaiduchek.maksym.api.exceptions;
+
+import gaiduchek.maksym.api.exceptions.exceptioncodes.ExceptionCodes;
+
+public class AccessException extends AbstractException {
+
+    public AccessException(ExceptionCodes code) {
+        super(code);
+    }
+}
