+++ api/src/main/java/gaiduchek/maksym/api/exceptions/AuthorizationException.java
@@ -1,0 +1,14 @@
+package gaiduchek.maksym.api.exceptions;
+
+import gaiduchek.maksym.api.exceptions.exceptioncodes.ExceptionCodes;
+
+public class AuthorizationException extends AbstractException {
+
+    public AuthorizationException(ExceptionCodes code) {
+        super(code);
+    }
+
+    public AuthorizationException(ExceptionCodes code, Object... formatArgs) {
+        super(code, formatArgs);
+    }
+}
