+++ api/src/main/java/gaiduchek/maksym/api/exceptions/EntityNotFoundException.java
@@ -1,0 +1,14 @@
+package gaiduchek.maksym.api.exceptions;
+
+import gaiduchek.maksym.api.exceptions.exceptioncodes.ExceptionCodes;
+
+public class EntityNotFoundException extends AbstractException {
+
+    public EntityNotFoundException(ExceptionCodes code) {
+        super(code);
+    }
+
+    public EntityNotFoundException(ExceptionCodes code, Object... formatArgs) {
+        super(code, formatArgs);
+    }
+}
