+++ api/src/main/java/gaiduchek/maksym/api/exceptions/exceptioncodes/ExceptionCodes.java
@@ -1,0 +1,8 @@
+package gaiduchek.maksym.api.exceptions.exceptioncodes;
+
+public interface ExceptionCodes {
+
+    String getCode();
+
+    String getDescription();
+}
