+++ api/src/main/java/gaiduchek/maksym/api/exceptions/exceptioncodes/ValidationExceptionCodes.java
@@ -1,0 +1,14 @@
+package gaiduchek.maksym.api.exceptions.exceptioncodes;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@RequiredArgsConstructor
+@Getter
+public enum ValidationExceptionCodes implements ExceptionCodes {
+
+    INVALID_DTO("SHOP-VLD-001", "Invalid DTO");
+
+    public final String code;
+    public final String description;
+}
