+++ api/src/main/java/gaiduchek/maksym/api/groups/CreateGroup.java
@@ -1,0 +1,5 @@
+package gaiduchek.maksym.api.groups;
+
+public interface CreateGroup {
+
+}
