+++ api/src/main/java/gaiduchek/maksym/api/groups/UsedInOtherGroup.java
@@ -1,0 +1,4 @@
+package gaiduchek.maksym.api.groups;
+
+public interface UsedInOtherGroup {
+}
