+++ api/src/main/java/gaiduchek/maksym/api/security/model/JwtAuthentication.java
@@ -1,0 +1,56 @@
+package gaiduchek.maksym.api.security.model;
+
+import gaiduchek.maksym.api.model.Role;
+import lombok.Getter;
+import lombok.Setter;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.GrantedAuthority;
+
+import java.util.Collection;
+import java.util.Set;
+
+@Getter
+@Setter
+public class JwtAuthentication implements Authentication {
+
+    private boolean authenticated;
+    private Long userId;
+    private String login;
+    private String firstName;
+    private Set<Role> roles;
+
+    @Override
+    public Collection<? extends GrantedAuthority> getAuthorities() {
+        return roles;
+    }
+
+    @Override
+    public Object getCredentials() {
+        return null;
+    }
+
+    @Override
+    public Object getDetails() {
+        return null;
+    }
+
+    @Override
+    public Object getPrincipal() {
+        return login;
+    }
+
+    @Override
+    public boolean isAuthenticated() {
+        return authenticated;
+    }
+
+    @Override
+    public void setAuthenticated(boolean isAuthenticated) throws IllegalArgumentException {
+        this.authenticated = isAuthenticated;
+    }
+
+    @Override
+    public String getName() {
+        return firstName;
+    }
+}
