+++ api/src/main/java/gaiduchek/maksym/api/security/services/interfaces/JwtProvider.java
@@ -1,0 +1,10 @@
+package gaiduchek.maksym.api.security.services.interfaces;
+
+import io.jsonwebtoken.Claims;
+
+public interface JwtProvider {
+
+    void validateAccessToken(String accessToken);
+
+    Claims getAccessClaims(String token);
+}
