+++ api/src/main/java/gaiduchek/maksym/api/utils/JwtUtils.java
@@ -1,0 +1,31 @@
+package gaiduchek.maksym.api.utils;
+
+import gaiduchek.maksym.api.model.Role;
+import gaiduchek.maksym.api.security.model.JwtAuthentication;
+import io.jsonwebtoken.Claims;
+import lombok.AccessLevel;
+import lombok.NoArgsConstructor;
+
+import java.util.List;
+import java.util.Set;
+import java.util.stream.Collectors;
+
+@NoArgsConstructor(access = AccessLevel.PRIVATE)
+public final class JwtUtils {
+
+    public static JwtAuthentication generate(Claims claims) {
+        final JwtAuthentication jwtInfoToken = new JwtAuthentication();
+        jwtInfoToken.setRoles(getRoles(claims));
+        jwtInfoToken.setFirstName(claims.get("firstName", String.class));
+        jwtInfoToken.setLogin(claims.getSubject());
+        jwtInfoToken.setUserId(claims.get("userId", Integer.class).longValue());
+        return jwtInfoToken;
+    }
+
+    private static Set<Role> getRoles(Claims claims) {
+        final List<String> roles = claims.get("roles", List.class);
+        return roles.stream()
+                .map(Role::valueOf)
+                .collect(Collectors.toSet());
+    }
+}
