--- src/main/java/com/maximgaiduchek/eshop/EShopApplication.java
@@ -1,12 +1,0 @@
-package com.maximgaiduchek.eshop;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class EShopApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(EShopApplication.class, args);
-    }
-}
