--- src/main/java/com/maximgaiduchek/eshop/configurations/PersistenceConfig.java
@@ -1,19 +1,0 @@
-package com.maximgaiduchek.eshop.configurations;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.data.auditing.DateTimeProvider;
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
-
-import java.time.ZonedDateTime;
-import java.util.Optional;
-
-@Configuration
-@EnableJpaAuditing(dateTimeProviderRef = "auditingDateTimeProvider")
-public class PersistenceConfig {
-
-    @Bean
-    public DateTimeProvider auditingDateTimeProvider() {
-        return () -> Optional.of(ZonedDateTime.now());
-    }
-}
