--- src/main/java/com/maximgaiduchek/eshop/exceptions/AccessException.java
@@ -1,11 +1,0 @@
-package com.maximgaiduchek.eshop.exceptions;
-
-
-import com.maximgaiduchek.eshop.exceptions.exceptioncodes.ExceptionCodes;
-
-public class AccessException extends AbstractException {
-
-    public AccessException(ExceptionCodes code) {
-        super(code);
-    }
-}
