--- src/main/java/com/maximgaiduchek/eshop/exceptions/exceptioncodes/ExceptionCodes.java
@@ -1,8 +1,0 @@
-package com.maximgaiduchek.eshop.exceptions.exceptioncodes;
-
-public interface ExceptionCodes {
-
-    String getCode();
-
-    String getDescription();
-}
