--- src/main/java/com/maximgaiduchek/eshop/security/services/interfaces/JwtProvider.java
@@ -1,10 +1,0 @@
-package com.maximgaiduchek.eshop.security.services.interfaces;
-
-import io.jsonwebtoken.Claims;
-
-public interface JwtProvider {
-
-    void validateAccessToken(String accessToken);
-
-    Claims getAccessClaims(String token);
-}
