--- src/main/java/com/maximgaiduchek/eshop/validation/groups/UpdateGroup.java
@@ -1,5 +1,0 @@
-package com.maximgaiduchek.eshop.validation.groups;
-
-public interface UpdateGroup {
-
-}
