--- src/main/java/com/maximgaiduchek/eshop/validation/groups/UsedInOtherGroup.java
@@ -1,4 +1,0 @@
-package com.maximgaiduchek.eshop.validation.groups;
-
-public interface UsedInOtherGroup {
-}
