--- src/main/resources/application.yml
@@ -1,62 +1,0 @@
-spring:
-  application:
-    name: 'api'
-  mvc:
-    dispatch-options-request: true
-  batch:
-    job:
-      enabled: false
-  main:
-    allow-bean-definition-overriding: true
-  servlet:
-    multipart:
-      enabled: true
-      file-size-threshold: '2KB'
-      max-file-size: '10MB'
-      max-request-size: '10MB'
-  datasource:
-    username: '${DB_USERNAME}'
-    password: '${DB_PASSWORD}'
-    url: 'jdbc:postgresql://${DB_ENDPOINT}/'
-    name: 'postgres'
-  jpa:
-    show-sql: false
-    properties:
-      hibernate:
-        show_sql: false
-        format_sql: true
-        dialect: 'org.hibernate.dialect.PostgreSQLDialect'
-        jdbc:
-          time_zone: 'UTC'
-    hibernate:
-      ddl-auto: 'update'
-#  flyway:
-#    baseline-on-migrate: true
-#    enabled: true
-#    schemas: 'public'
-#    validate-on-migrate: false
-#    user: '${DB_USERNAME}'
-#    password: '${DB_PASSWORD}'
-#    url: 'jdbc:postgresql://${DB_ENDPOINT}/'
-#    out-of-order: true
-
-server:
-  error:
-    include-message: always
-    include-binding-errors: always
-  port: 8080
-  servlet:
-    context-path: '/api'
-
-service:
-  domain-url: 'https://${DOMAIN}/'
-
-jwt:
-  secret:
-    access: '${JWT_ACCESS_KEY}'
-    refresh: '${JWT_REFRESH_KEY}'
-
-logging:
-  level:
-    com.maximgaiduchek: TRACE
-    # org.springframework: TRACE
