--- src/test/java/com/maximgaiduchek/eshop/EShopApplicationTests.java
@@ -1,13 +1,0 @@
-package com.maximgaiduchek.eshop;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class EShopApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
